Separate bot onto modules
The codebase was pretty weird due to the fact that everything was being done from one file (server.js)

I separated the code, now it has 4 files and the responsibilities are a bit clearer. Also, I tried really hard to work with promises wherever I could. I don't know if that approach is the best it could be, but it was a good way for me to understand how promises work. It has been a good journey try and make these perform as well as the "only one file containing everything" approach and I feel like I learned a lot.

Also, I did some refactoring on the way everything is written. More arrow functions (and a bit more consistency on how these are presented), camel case in all methods... Oh and I also removed all semicolons. Seems prettier this way!